Clean up TodaysHabitsFragment + AllHabitsFragment
Created an abstract HabitListFragment that is extended by both since
its logic only differs in the habit list filtering. This can also be
used for viewing another user's public habits in the future by removing
the floating action button.
