+++ app/src/main/java/com/github/cmput301f21t44/hellohabits/view/MainActivity.java
@@ -1,9 +1,11 @@
 package com.github.cmput301f21t44.hellohabits.view;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.navigation.NavController;
 import androidx.navigation.Navigation;
@@ -14,12 +16,14 @@ import com.github.cmput301f21t44.hellohabits.R;
 import com.github.cmput301f21t44.hellohabits.databinding.ActivityMainBinding;
 import com.github.cmput301f21t44.hellohabits.viewmodel.PreviousListViewModel;
 import com.github.cmput301f21t44.hellohabits.viewmodel.ViewModelFactory;
+import com.google.firebase.auth.FirebaseAuth;
 
 public class MainActivity extends AppCompatActivity {
     private AppBarConfiguration mAppBarConfiguration;
     private NavController mNavController;
     private PreviousListViewModel mPreviousListViewModel;
     private boolean mFromViewHabit;
+    private FirebaseAuth mAuth;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -29,6 +33,7 @@ public class MainActivity extends AppCompatActivity {
         setContentView(binding.getRoot());
 
         setSupportActionBar(binding.appBarMain.toolbar);
+        mAuth = FirebaseAuth.getInstance();
 
         mPreviousListViewModel = ViewModelFactory.getProvider(this)
                 .get(PreviousListViewModel.class);
@@ -48,6 +53,16 @@ public class MainActivity extends AppCompatActivity {
 
         mPreviousListViewModel.getFromViewHabit()
                 .observe(this, val -> this.mFromViewHabit = val);
+        requireUser();
+    }
+
+    /**
+     * Navigates to login if there is no user
+     */
+    private void requireUser() {
+        if (mAuth.getCurrentUser() == null) {
+            mNavController.navigate(R.id.loginFragment);
+        }
     }
 
     @Override
@@ -64,7 +79,10 @@ public class MainActivity extends AppCompatActivity {
         // as you specify a parent activity in AndroidManifest.xml.
         int id = item.getItemId();
 
-        if (id == R.id.action_settings) return true;
+        if (id == R.id.action_sign_out) {
+            showSignOutDialog();
+            return true;
+        }
 
         return super.onOptionsItemSelected(item);
     }
@@ -79,5 +97,27 @@ public class MainActivity extends AppCompatActivity {
 
         return NavigationUI.navigateUp(mNavController, mAppBarConfiguration)
                 || super.onSupportNavigateUp();
+    }
+
+    /**
+     * Show an AlertDialog when trying to sign out
+     */
+    private void showSignOutDialog() {
+        new AlertDialog.Builder(this)
+                .setTitle("Sign out")
+                .setMessage("Are you sure you want to sign out?")
+                .setIcon(R.drawable.ic_baseline_warning_24)
+                .setPositiveButton("YES", (dialog, b) -> signOut())
+                .setNegativeButton("NO", null).show();
+    }
+
+    /**
+     * Signs out and restarts MainActivity to clear all LiveData listeners
+     */
+    private void signOut() {
+        mAuth.signOut();
+        Intent intent = new Intent(this, MainActivity.class);
+        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
+        startActivity(intent);
     }
 }
