--- app/src/main/res/layout/fragment_todays_habits.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".view.habit.TodaysHabitsFragment">
-
-    <Button
-        android:id="@+id/button_new_habit"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="@dimen/fragment_horizontal_margin"
-        android:text="@string/new_habit"
-        app:layout_constraintBottom_toTopOf="@+id/social"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/view_all_habits" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/habit_recycler_view"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:layout_marginHorizontal="@dimen/fragment_horizontal_margin"
-        android:layout_marginVertical="@dimen/fragment_vertical_margin"
-        app:layout_constraintBottom_toTopOf="@+id/view_all_habits"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="1.0"
-        tools:itemCount="5"
-        tools:listitem="@layout/list_habit_item" />
-
-    <Button
-        android:id="@+id/social"
-        android:layout_width="match_parent"
-        android:layout_height="48dp"
-        android:layout_marginHorizontal="@dimen/fragment_horizontal_margin"
-        android:layout_marginBottom="@dimen/fragment_vertical_margin"
-        android:text="@string/sign_out"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/button_new_habit" />
-
-    <Button
-        android:id="@+id/view_all_habits"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="@dimen/fragment_horizontal_margin"
-        android:text="@string/view_all_habits"
-        app:layout_constraintBottom_toTopOf="@+id/button_new_habit"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/habit_recycler_view" />
-</androidx.constraintlayout.widget.ConstraintLayout>
