+++ app/src/main/res/menu/menu_main.xml
@@ -3,8 +3,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     tools:context="com.github.cmput301f21t44.hellohabits.view.MainActivity">
     <item
-        android:id="@+id/action_settings"
+        android:id="@+id/action_sign_out"
         android:orderInCategory="100"
-        android:title="@string/action_settings"
+        android:title="@string/sign_out"
         app:showAsAction="never" />
 </menu>
